Crate nebari

Transactional append-only B-Tree storage for BonsaiDb.

Modules

io

IO abstractions for Nebari.

transaction

ACID-compliant transaction log and manager.

tree

Append-only B-Tree implementation

Structs

Buffer

A wrapper around a Cow<'a, [u8]> wrapper that implements Read, and has a convenience method to take a slice of bytes as another Buffer which shares a reference to the same underlying Cow.

ChunkCache

A configurable cache that operates at the “chunk” level.

Config

A database configuration used to open a database.

Context

A shared environment for database operations.

Error

An error from Nebari as well as an associated backtrace.

ExecutingTransaction

An executing transaction. While this exists, no other transactions can execute across the same trees as this transaction holds.

Roots

A multi-tree transactional B-Tree database.

ThreadPool

A thread pool that commits transactions to disk in parallel.

TransactionTree

A tree that is modifiable during a transaction.

Tree

A named collection of keys and values.

Enums

AbortError

An error that could come from user code or Roots.

CompareAndSwapError

An error returned from compare_and_swap().

ErrorKind

An error from Nebari.

Traits

Vault

A provider of encryption for blocks of data.