Module nebari::io

IO abstractions for Nebari.

Nebari was written to have the flexibility of multiple backend options. This may allow Nebari to target no_std in the future or allow for other IO strategies to be implemented in addition to the ones seen here today.

Modules

fs

Filesystem IO provided by std::fs.

memory

A virtual memory-based filesystem.

Structs

PathIds

Converts between paths and unique IDs.

RecreatedFile

A file that has had its contents replaced. While this value exists, all other threads will be blocked from interacting with the PathIds structure. Only hold onto this value for short periods of time.

Traits

FileManager

A manager that is responsible for controlling write access to a file.

FileOp

An operation to perform on a file.

ManagedFile

A file that can be interacted with using async operations.

OpenableFile

A file that can have operations performed on it.