Struct nebari::io::memory::MemoryFile

pub struct MemoryFile { /* fields omitted */ }

A fake “file” represented by an in-memory buffer. This should only be used in testing, as this database format is not optimized for memory efficiency.