Trait nebari::io::File

pub trait File: Debug + Send + Sync + Seek + Read + Write + 'static {
    fn id(&self) -> Option<u64>;

    fn path(&self) -> &Path;

    fn length(&self) -> Result<u64, Error>;

    fn close(self) -> Result<(), Error>;
}

A generic file trait.

Required methods

fn id(&self) -> Option<u64>

Returns the unique ID of this file. Only unique within the manager it was opened from.

fn path(&self) -> &Path

Returns the path to the file.

fn length(&self) -> Result<u64, Error>

Returns the length of the file.

fn close(self) -> Result<(), Error>

Safely closes the file after flushing any pending operations to disk.