Trait nebari::io::FileManager

pub trait FileManager: Send + Sync + Clone + Default + Debug + 'static {
    type File: ManagedFile<Manager = Self>;
    type FileHandle: OpenableFile<Self::File>;
    fn read(&self, path: impl AsRef<Path>) -> Result<Self::FileHandle, Error>;

    fn append(&self, path: impl AsRef<Path>) -> Result<Self::FileHandle, Error>;

    fn close_handles<F: FnOnce(u64)>(
        &self, 
        path: impl AsRef<Path>, 
        publish_callback: F
    );

    fn delete(&self, path: impl AsRef<Path>) -> Result<bool, Error>;

    fn delete_directory(&self, path: impl AsRef<Path>) -> Result<(), Error>;

    fn file_length(&self, path: impl AsRef<Path>) -> Result<u64, Error> { ... }

    fn exists(&self, path: impl AsRef<Path>) -> Result<bool, Error> { ... }
}

A manager that is responsible for controlling write access to a file.

Associated Types

type File: ManagedFile<Manager = Self>

The ManagedFile that this manager is for.

type FileHandle: OpenableFile<Self::File>

A file handle type, which can have operations executed against it.

Required methods

fn read(&self, path: impl AsRef<Path>) -> Result<Self::FileHandle, Error>

Returns a file handle that can be used for reading operations.

fn append(&self, path: impl AsRef<Path>) -> Result<Self::FileHandle, Error>

Returns a file handle that can be used to read and write.

fn close_handles<F: FnOnce(u64)>(
&self,
path: impl AsRef<Path>,
publish_callback: F
)

Closes all open handles for path, and calls publish_callback before unlocking any locks aquired during the operation.

fn delete(&self, path: impl AsRef<Path>) -> Result<bool, Error>

Check if the file exists.

fn delete_directory(&self, path: impl AsRef<Path>) -> Result<(), Error>

Removes a directory and all of its contents.

Provided methods

fn file_length(&self, path: impl AsRef<Path>) -> Result<u64, Error>

Returns the length of the file.

fn exists(&self, path: impl AsRef<Path>) -> Result<bool, Error>

Check if the file exists.