Trait nebari::io::ManagedFile

pub trait ManagedFile: Send + Sync + Seek + Read + Write + Sized + 'static {
    type Manager: FileManager<File = Self>;
    fn id(&self) -> Option<u64>;

    fn path(&self) -> &Path;

    fn open_for_read(
        path: impl AsRef<Path> + Send, 
        id: Option<u64>
    ) -> Result<Self, Error>;

    fn open_for_append(
        path: impl AsRef<Path> + Send, 
        id: Option<u64>
    ) -> Result<Self, Error>;

    fn length(&self) -> Result<u64, Error>;

    fn close(self) -> Result<(), Error>;
}

A file that can be interacted with using async operations.

This trait is an abstraction that mimics tokio-uring’s File type, allowing for a non-uring implementation to be provided as well. This is why the read/write APIs take ownership of the buffer – to satisfy the requirements of tokio-uring.

Associated Types

type Manager: FileManager<File = Self>

The file manager that synchronizes file access across threads.

Required methods

fn id(&self) -> Option<u64>

Returns the unique ID of this file. Only unique within Self::Manager.

fn path(&self) -> &Path

Returns the path to the file.

fn open_for_read(
path: impl AsRef<Path> + Send,
id: Option<u64>
) -> Result<Self, Error>

Opens a file at path with read-only permission.

fn open_for_append(
path: impl AsRef<Path> + Send,
id: Option<u64>
) -> Result<Self, Error>

Opens or creates a file at path, positioning the cursor at the end of the file.

fn length(&self) -> Result<u64, Error>

Returns the length of the file.

fn close(self) -> Result<(), Error>

Safely closes the file after flushing any pending operations to disk.