Trait nebari::io::OpenableFile

pub trait OpenableFile<F: ManagedFile>: Debug + Sized + Send + Sync {
    fn id(&self) -> Option<u64>;

    fn execute<W: FileOp<F>>(&mut self, operator: W) -> W::Output;

    fn replace_with<C: FnOnce(u64)>(
        self, 
        path: &Path, 
        manager: &F::Manager, 
        publish_callback: C
    ) -> Result<Self, Error>;

    fn close(self) -> Result<(), Error>;
}

A file that can have operations performed on it.

Required methods

fn id(&self) -> Option<u64>

Returns the id of the file assigned from the file manager.

fn execute<W: FileOp<F>>(&mut self, operator: W) -> W::Output

Executes an operation.

fn replace_with<C: FnOnce(u64)>(
self,
path: &Path,
manager: &F::Manager,
publish_callback: C
) -> Result<Self, Error>

Replaces the current file with the file located at path atomically.

fn close(self) -> Result<(), Error>

Closes the file. This may not actually close the underlying file, depending on what other tasks have access to the underlying file as well.