Struct nebari::Buffer

pub struct Buffer<'a> { /* fields omitted */ }

A wrapper around a Cow<'a, [u8]> wrapper that implements Read, and has a convenience method to take a slice of bytes as another Buffer which shares a reference to the same underlying Cow.

Implementations

impl<'a> Buffer<'a>

pub fn to_owned(&self) -> Buffer<'static>

Converts this buffer into its slice and returns a static-lifetimed instance.

pub fn read_bytes(&mut self, count: usize) -> Result<Self, Error>

Reads count bytes from the front of the buffer, returning a buffer that shares the same underlying buffer.

pub fn as_slice(&self) -> &[u8]

Returns this buffer as a slice.