Struct nebari::TransactionTree

pub struct TransactionTree<Root: Root, F: ManagedFile> { /* fields omitted */ }

A tree that is modifiable during a transaction.

Implementations

impl<F: ManagedFile> TransactionTree<VersionedTreeRoot, F>

pub fn current_sequence_id(&self) -> u64

Returns the latest sequence id.

impl<Root: Root, F: ManagedFile> TransactionTree<Root, F>

pub fn set(
&mut self,
key: impl Into<Buffer<'static>>,
value: impl Into<Buffer<'static>>
) -> Result<(), Error>

Sets key to value.

pub fn replace(
&mut self,
key: impl Into<Buffer<'static>>,
value: impl Into<Buffer<'static>>
) -> Result<Option<Buffer<'static>>, Error>

Sets key to value. If a value already exists, it will be returned.

pub fn get(&mut self, key: &[u8]) -> Result<Option<Buffer<'static>>, Error>

Returns the current value of key. This will return updated information if it has been previously updated within this transaction.

pub fn remove(&mut self, key: &[u8]) -> Result<Option<Buffer<'static>>, Error>

Removes key and returns the existing value, if present.

pub fn compare_and_swap(
&mut self,
key: &[u8],
old: Option<&Buffer<'_>>,
new: Option<Buffer<'_>>
) -> Result<(), CompareAndSwapError>

Compares the value of key against old. If the values match, key will be set to the new value if new is Some or removed if new is None.

pub fn get_multiple(
&mut self,
keys: &[&[u8]]
) -> Result<Vec<(Buffer<'static>, Buffer<'static>)>, Error>

Retrieves the values of keys. If any keys are not found, they will be omitted from the results.

pub fn get_range<'b, B: RangeBounds<Buffer<'b>> + Debug + 'static>(
&mut self,
range: B
) -> Result<Vec<(Buffer<'static>, Buffer<'static>)>, Error>

Retrieves all of the values of keys within range.

pub fn scan<'b, E, B, KeyEvaluator, DataCallback>(
&mut self,
range: B,
forwards: bool,
key_evaluator: KeyEvaluator,
callback: DataCallback
) -> Result<(), AbortError<E>> where
B: RangeBounds<Buffer<'b>> + Debug + 'static,
KeyEvaluator: FnMut(&Buffer<'static>) -> KeyEvaluation,
DataCallback: FnMut(Buffer<'static>, Buffer<'static>) -> Result<(), AbortError<E>>,
E: Display + Debug,

Scans the tree. Each key that is contained range will be passed to key_evaluator, which can opt to read the data for the key, skip, or stop scanning. If KeyEvaluation::ReadData is returned, callback will be invoked with the key and stored value. The order in which callback is invoked is not necessarily the same order in which the keys are found.

pub fn last_key(&mut self) -> Result<Option<Buffer<'static>>, Error>

Returns the last of the tree.

pub fn last(
&mut self
) -> Result<Option<(Buffer<'static>, Buffer<'static>)>, Error>

Returns the last key and value of the tree.