Struct nebari::Tree

pub struct Tree<Root: Root, File: ManagedFile> { /* fields omitted */ }

A named collection of keys and values.

Implementations

impl<Root: Root, File: ManagedFile> Tree<Root, File>

pub fn name(&self) -> &str

Returns the name of the tree.

pub fn path(&self) -> PathBuf

Returns the path to the file for this tree.

pub fn count(&self) -> u64

Returns the number of keys stored in the tree. Does not include deleted keys.

pub fn set(
&self,
key: impl Into<Buffer<'static>>,
value: impl Into<Buffer<'static>>
) -> Result<(), Error>

Sets key to value. This is executed within its own transaction.

pub fn get(&self, key: &[u8]) -> Result<Option<Buffer<'static>>, Error>

Retrieves the current value of key, if present. Does not reflect any changes in pending transactions.

pub fn remove(&self, key: &[u8]) -> Result<Option<Buffer<'static>>, Error>

Removes key and returns the existing value, if present. This is executed within its own transaction.

pub fn compare_and_swap(
&self,
key: &[u8],
old: Option<&Buffer<'_>>,
new: Option<Buffer<'_>>
) -> Result<(), CompareAndSwapError>

Compares the value of key against old. If the values match, key will be set to the new value if new is Some or removed if new is None. This is executed within its own transaction.

pub fn get_multiple(
&self,
keys: &[&[u8]]
) -> Result<Vec<(Buffer<'static>, Buffer<'static>)>, Error>

Retrieves the values of keys. If any keys are not found, they will be omitted from the results.

pub fn get_range<'range, Range: RangeBounds<Buffer<'range>> + Clone + Debug + 'static>(
&self,
range: Range
) -> Result<Vec<(Buffer<'static>, Buffer<'static>)>, Error>

Retrieves all of the values of keys within range.

pub fn scan<'range, CallerError, Range, NodeEvaluator, KeyEvaluator, DataCallback>(
&self,
range: Range,
forwards: bool,
node_evaluator: NodeEvaluator,
key_evaluator: KeyEvaluator,
callback: DataCallback
) -> Result<(), AbortError<CallerError>> where
Range: RangeBounds<Buffer<'range>> + Clone + Debug + 'static,
NodeEvaluator: FnMut(&Buffer<'static>, &Root::ReducedIndex, usize) -> bool,
KeyEvaluator: FnMut(&Buffer<'static>, &Root::Index) -> KeyEvaluation,
DataCallback: FnMut(Buffer<'static>, &Root::Index, Buffer<'static>) -> Result<(), AbortError<CallerError>>,
CallerError: Display + Debug,

Scans the tree. Each key that is contained range will be passed to key_evaluator, which can opt to read the data for the key, skip, or stop scanning. If KeyEvaluation::ReadData is returned, callback will be invoked with the key and stored value. The order in which callback is invoked is not necessarily the same order in which the keys are found.

pub fn last_key(&self) -> Result<Option<Buffer<'static>>, Error>

Returns the last key of the tree.

pub fn last(&self) -> Result<Option<(Buffer<'static>, Buffer<'static>)>, Error>

Returns the last key and value of the tree.

pub fn compact(&self) -> Result<(), Error>

Rewrites the database to remove data that is no longer current. Because Nebari uses an append-only format, this is helpful in reducing disk usage.

See TreeFile::compact() for more information.

impl<File: ManagedFile, Index> Tree<VersionedTreeRoot<Index>, File> where
Index: EmbeddedIndex + Reducer<Index> + Clone + Debug + 'static,

pub fn scan_sequences<CallerError, Range, KeyEvaluator, DataCallback>(
&mut self,
range: Range,
forwards: bool,
key_evaluator: &mut KeyEvaluator,
data_callback: &mut DataCallback
) -> Result<(), AbortError<CallerError>> where
Range: Clone + RangeBounds<u64> + Debug + 'static,
KeyEvaluator: FnMut(KeySequence) -> KeyEvaluation,
DataCallback: FnMut(KeySequence, Buffer<'static>) -> Result<(), AbortError<CallerError>>,
CallerError: Display + Debug,

Scans the tree for keys that are contained within range. If forwards is true, scanning starts at the lowest sort-order key and scans forward. Otherwise, scanning starts at the highest sort-order key and scans backwards. key_evaluator is invoked for each key as it is encountered. For all KeyEvaluation::ReadData results returned, callback will be invoked with the key and values. The callback may not be invoked in the same order as the keys are scanned.