Trait nebari::FileManager

pub trait FileManager: Send + Sync + Clone + Default + Debug + 'static {
    type File: ManagedFile<Manager = Self>;
    type FileHandle: OpenableFile<Self::File>;
    fn read(
        &self, 
        path: impl AsRef<Path> + Send
    ) -> Result<Self::FileHandle, Error>;

    fn append(
        &self, 
        path: impl AsRef<Path> + Send
    ) -> Result<Self::FileHandle, Error>;

    fn delete(&self, path: impl AsRef<Path> + Send) -> Result<bool, Error>;

    fn file_length(&self, path: impl AsRef<Path> + Send) -> Result<u64, Error> { ... }

    fn exists(&self, path: impl AsRef<Path> + Send) -> Result<bool, Error> { ... }
}

A manager that is responsible for controlling write access to a file.

Associated Types

type File: ManagedFile<Manager = Self>

The ManagedFile that this manager is for.

type FileHandle: OpenableFile<Self::File>

A file handle type, which can have operations executed against it.

Required methods

fn read(&self, path: impl AsRef<Path> + Send) -> Result<Self::FileHandle, Error>

Returns a file handle that can be used for reading operations.

fn append(
&self,
path: impl AsRef<Path> + Send
) -> Result<Self::FileHandle, Error>

Returns a file handle that can be used to read and write.

fn delete(&self, path: impl AsRef<Path> + Send) -> Result<bool, Error>

Check if the file exists.

Provided methods

fn file_length(&self, path: impl AsRef<Path> + Send) -> Result<u64, Error>

Returns the length of the file.

fn exists(&self, path: impl AsRef<Path> + Send) -> Result<bool, Error>

Check if the file exists.