Trait nebari::Vault

pub trait Vault: Debug + Send + Sync + 'static {
    type Error: SendSyncError;
    fn encrypt(&self, payload: &[u8]) -> Result<Vec<u8>, Self::Error>;

    fn decrypt(&self, payload: &[u8]) -> Result<Vec<u8>, Self::Error>;
}

A provider of encryption for blocks of data.

Associated Types

type Error: SendSyncError

The error type that the vault can produce.

Required methods

fn encrypt(&self, payload: &[u8]) -> Result<Vec<u8>, Self::Error>

Encrypts payload, returning a new buffer that contains all information necessary to decrypt it in the future.

fn decrypt(&self, payload: &[u8]) -> Result<Vec<u8>, Self::Error>

Decrypts a previously encrypted payload, returning the decrypted information.