Module nebari::transaction

ACID-compliant transaction log and manager.

Structs

Entry

A single record of a change.

LogEntry

An entry in a transaction log.

State

The transaction log state.

TransactionHandle

A handle to an executing transaction.

TransactionLog

A transaction log that records changes for one or more trees.

TransactionManager

A shared TransactionLog manager. Allows multiple threads to interact with a single transaction log.

TreeLock

An acquirable lock for a tree.

TreeLockHandle

A handle to a lock. Upon dropping, the lock will be released.

Enums

Entries

A collection of entries that might be borrowed.

Type Definitions

TransactionChanges

A map of entries by their tree names.

TreeLocks

A collection of handles that keep trees locked.