Struct nebari::tree::TreeFile

pub struct TreeFile<Root: Root, F: ManagedFile> {
    pub state: State<Root>,
    // some fields omitted
}

An append-only tree file.

Generics

F: An ManagedFile implementor.

MAX_ORDER: The maximum number of children a node in the tree can contain. This implementation attempts to grow naturally towards this upper limit. Changing this parameter does not automatically rebalance the tree, but over time the tree will be updated.

Fields
state: State<Root>

The state of the file.

Implementations

impl<Root: Root, F: ManagedFile> TreeFile<Root, F>

pub fn new(
file: <F::Manager as FileManager>::FileHandle,
state: State<Root>,
vault: Option<Arc<dyn Vault>>,
cache: Option<ChunkCache>
) -> Result<Self, Error>

Returns a tree as contained in file.

state should already be initialized using Self::initialize_state if the file exists.

pub fn read(
path: impl AsRef<Path>,
state: State<Root>,
context: &Context<F::Manager>,
transactions: Option<&TransactionManager<F::Manager>>
) -> Result<Self, Error>

Opens a tree file with read-only permissions.

pub fn write(
path: impl AsRef<Path>,
state: State<Root>,
context: &Context<F::Manager>,
transactions: Option<&TransactionManager<F::Manager>>
) -> Result<Self, Error>

Opens a tree file with the ability to read and write.

pub fn initialize_state(
state: &State<Root>,
file_path: &Path,
context: &Context<F::Manager>,
transaction_manager: Option<&TransactionManager<F::Manager>>
) -> Result<(), Error>

Attempts to load the last saved state of this tree into state.

pub fn push(
&mut self,
transaction_id: u64,
key: Buffer<'static>,
document: Buffer<'static>
) -> Result<(), Error>

Pushes a key/value pair. Replaces any previous value if set.

pub fn modify(
&mut self,
modification: Modification<'_, Buffer<'static>>
) -> Result<(), Error>

Executes a modification.

pub fn get<'k>(
&mut self,
key: &'k [u8],
in_transaction: bool
) -> Result<Option<Buffer<'static>>, Error>

Gets the value stored for key.

pub fn get_multiple(
&mut self,
keys: &[&[u8]],
in_transaction: bool
) -> Result<Vec<(Buffer<'static>, Buffer<'static>)>, Error>

Gets the value stored for key.

pub fn get_range<'b, B: RangeBounds<Buffer<'b>> + Debug + 'static>(
&mut self,
range: B,
in_transaction: bool
) -> Result<Vec<(Buffer<'static>, Buffer<'static>)>, Error>

Gets the value stored for key.

pub fn scan<'b, E, B, KeyEvaluator, DataCallback>(
&mut self,
range: B,
forwards: bool,
in_transaction: bool,
key_evaluator: KeyEvaluator,
callback: DataCallback
) -> Result<(), AbortError<E>> where
B: RangeBounds<Buffer<'b>> + Debug + 'static,
KeyEvaluator: FnMut(&Buffer<'static>) -> KeyEvaluation,
DataCallback: FnMut(Buffer<'static>, Buffer<'static>) -> Result<(), AbortError<E>>,
E: Display + Debug,

Gets the value stored for key.

pub fn commit(&mut self) -> Result<(), Error>

Commits the tree. This is only needed if writes were done with a transaction id. This will fully flush the tree and publish the transactional state to be available to readers.