Trait nebari::tree::EmbeddedIndex

pub trait EmbeddedIndex: Serializable + Send + Sync + Sized + 'static {
    fn index(key: &Buffer<'_>, value: Option<&Buffer<'static>>) -> Self;
}

An index that is embeddable within a tree.

An index is a computed value that is stored directly within the B-Tree structure. Because these are encoded directly onto the nodes, they should be kept shorter for better performance.

Required methods

fn index(key: &Buffer<'_>, value: Option<&Buffer<'static>>) -> Self

Index the key and value.