Trait nebari::tree::Reducer

pub trait Reducer<Index> {
    fn reduce(indexes: &[&Index]) -> Self;

    fn rereduce(reductions: &[&Self]) -> Self;
}

Reduces one or more Indexes or instances of Self into a single Self value.

This trait is used to collect statistics within the B-Tree. The Index value is stored at the KeyEntry level, and each Interior entry contains Self, which is calculated by calling Reducer::reduce() on all stored Indexes.

When an Interior node points to other interior nodes, it calculates the stored value by calling Reducer::rereduce() with all the stored Self values.

Required methods

fn reduce(indexes: &[&Index]) -> Self

Reduces one or more indexes into a single reduced index.

fn rereduce(reductions: &[&Self]) -> Self

Reduces one or more previously-reduced indexes into a single reduced index.