Trait nebari::tree::Serializable

pub trait Serializable: Send + Sync + Sized + 'static {
    fn serialize_to<W: WriteBytesExt>(
        &self, 
        writer: &mut W
    ) -> Result<usize, Error>;

    fn deserialize_from<R: ReadBytesExt>(reader: &mut R) -> Result<Self, Error>;
}

A type that can be serialized and deserialized.

Required methods

fn serialize_to<W: WriteBytesExt>(&self, writer: &mut W) -> Result<usize, Error>

Serializes into writer and returns the number of bytes written.

fn deserialize_from<R: ReadBytesExt>(reader: &mut R) -> Result<Self, Error>

Deserializes from reader, and returns the deserialized instance. Implementors should not expect for the reader to be fully consumed at the end of this call.