Struct nebari::transaction::TransactionLog

pub struct TransactionLog<File: ManagedFile> { /* fields omitted */ }

A transaction log that records changes for one or more trees.

Implementations

impl<File: ManagedFile> TransactionLog<File>

pub fn read(
log_path: &Path,
state: State,
context: Context<File::Manager>
) -> Result<Self, Error>

Opens a transaction log for reading.

pub fn open(
log_path: &Path,
state: State,
context: Context<File::Manager>
) -> Result<Self, Error>

Opens a transaction log for writing.

pub fn total_size(&self) -> u64

Returns the total size of the transaction log file.

pub fn initialize_state(
state: &State,
context: &Context<File::Manager>
) -> Result<(), Error>

Initializes state to contain the information about the transaction log located at log_path.

pub fn push(&mut self, handles: Vec<LogEntry<'static>>) -> Result<(), Error>

Logs one or more transactions. After this call returns, the transaction log is guaranteed to be fully written to disk.

pub fn get(&mut self, id: u64) -> Result<Option<LogEntry<'static>>, Error>

Returns the executed transaction with the id provided. Returns None if not found.

pub fn scan<Callback: FnMut(LogEntry<'static>) -> bool>(
&mut self,
ids: impl RangeBounds<u64>,
callback: Callback
) -> Result<(), Error>

Logs one or more transactions. After this call returns, the transaction log is guaranteed to be fully written to disk.

pub fn close(self) -> Result<(), Error>

Closes the transaction log.

pub fn current_transaction_id(&self) -> u64

Returns the current transaction id.

pub fn new_transaction<'a, I: IntoIterator<Item = &'a [u8], IntoIter = II>, II: ExactSizeIterator<Item = &'a [u8]>>(
&self,
trees: I
) -> TransactionHandle

Begins a new transaction, exclusively locking trees.

pub fn state(&self) -> State

Returns the current state of the log.