Module nebari::tree

Append-only B-Tree implementation

The file format is inspired by Couchstore. The main difference is that the file header has extra information to allow for cross-tree transactions.

Numbers and Alignment

All numbers are encoded in big-endian format/network byte order.

All values are tightly packed. There is no padding or alignment that isn’t explicitly included.

File pages

The file is written in pages that are 4,096 bytes long. Each page has single u8 representing the PageHeader. If data needs to span more than one page, every 4,096 byte boundary must contain a PageHeader::Continuation.

File Headers

If the header is a PageHeader::Header, the contents of the block will be a single chunk that contains a serialized BTreeRoot.

Chunks

Each time a document, B-Tree node, or header is written, it is written as a chunk. If a Vault is in-use, each chunk will be pre-processed by the vault before a CRC-32-BZIP2 checksum is calculated. A chunk is limited to 4 gigabytes of data (2^32).

The chunk is written as:

u32 - Data length, excluding the header.

u32 - CRC

[u8] - Contents

A data block may contain more than one chunk.

Structs

CompareSwap

A wrapper for a CompareSwapFn.

Modification

A tree modification.

PagedWriter

Writes data in pages, allowing for quick scanning through the file.

State

The current state of a tree file. Must be initialized before passing to TreeFile::new if the file already exists.

TreeFile

An append-only tree file.

TreeRoot

A named tree root.

UnversionedTreeRoot

A versioned B-Tree root. This tree root internally uses two btrees, one to keep track of all writes using a unique “sequence” ID, and one that keeps track of all key-value pairs.

VersionedTreeRoot

A versioned B-Tree root. This tree root internally uses two btrees, one to keep track of all writes using a unique “sequence” ID, and one that keeps track of all key-value pairs.

Enums

KeyEvaluation

The result of evaluating a key that was scanned.

KeyOperation

An operation to perform on a key.

KeyRange

One or more keys.

Operation

An operation that is performed on a set of keys.

Traits

Root

A B-Tree root implementation.

Type Definitions

CompareSwapFn

A function that is allowed to check the current value of a key and determine how to operate on it. The first parameter is the key, and the second parameter is the current value, if present.