Module nebari::tree

Append-only B-Tree implementation

The file format is loosely inspired by Couchstore. Nebari is not compatible with Couchstore in any way.

Numbers and Alignment

All numbers are encoded in big-endian format/network byte order.

All values are tightly packed. There is no padding or alignment that isn’t explicitly included.

File pages

The file is written in pages that are 4,096 bytes long. Each page has single u8 representing the PageHeader. If data needs to span more than one page, every PAGE_SIZE byte boundary must contain a PageHeader::Continuation.

File Headers

If the header is a PageHeader::Header, the contents of the block will be a single chunk that contains a serialized BTreeRoot.

Chunks

Each time a value, B-Tree node, or header is written, it is written as a chunk. If a Vault is in-use, each chunk will be pre-processed by the vault before a CRC-32-BZIP2 checksum is calculated. A chunk is limited to 4 gigabytes of data (2^32).

The chunk is written as:

u32 - Data length, excluding the header.

u32 - CRC

[u8] - Contents

A data block may contain more than one chunk.

Structs

ActiveState

An active state for a tree file.

BTreeEntry

A B-Tree entry that stores a list of key-Index pairs.

ByIdStats

The reduced index of both VersionedByIdIndex and UnversionedByIdIndex

BySequenceIndex

The index stored within VersionedTreeRoot::by_sequence_root.

BySequenceStats

The reduced index of BySequenceIndex.

CompareSwap

A wrapper for a CompareSwapFn.

Interior

An interior B-Tree node. Does not contain values directly, and instead points to a node located on-disk elsewhere.

KeyEntry

An entry for a key. Stores a single index value for a single key.

KeyRange

One or more keys.

KeySequence

A stored revision of a key.

Modification

A tree modification.

PagedWriter

Writes data in pages, allowing for quick scanning through the file.

State

The current state of a tree file. Must be initialized before passing to TreeFile::new if the file already exists.

TransactableCompaction

A compaction process that runs in concert with a transaction manager.

TreeFile

An append-only tree file.

TreeRoot

A named tree with a specific root type.

UnversionedByIdIndex

The index stored within UnversionedTreeRoot::by_id_root.

UnversionedTreeRoot

A versioned B-Tree root. This tree root internally uses two btrees, one to keep track of all writes using a unique “sequence” ID, and one that keeps track of all key-value pairs.

VersionedByIdIndex

The index stored within VersionedTreeRoot::by_id_root.

VersionedTreeRoot

A versioned B-Tree root. This tree root internally uses two btrees, one to keep track of all writes using a unique “sequence” ID, and one that keeps track of all key-value pairs.

Enums

BTreeNode

A B-Tree entry that stores a list of key-Index pairs.

KeyEvaluation

The result of evaluating a key that was scanned.

KeyOperation

An operation to perform on a key.

Operation

An operation that is performed on a set of keys.

PageHeader

The header byte for a tree file’s page.

Pointer

A pointer to a location on-disk. May also contain the node already loaded.

Constants

PAGE_SIZE

The number of bytes in each page on-disk.

Traits

AnyTreeRoot

A named tree that can be used in a transaction.

EmbeddedIndex

An index that is embeddable within a tree.

Reducer

Reduces one or more Indexes or instances of Self into a single Self value.

Root

A B-Tree root implementation.

Serializable

A type that can be serialized and deserialized.

ValueIndex

An index that serializes a value to the file.

Type Definitions

CompareSwapFn

A function that is allowed to check the current value of a key and determine how to operate on it. The first parameter is the key, and the second parameter is the current value, if present.

Unversioned

An unversioned tree with no additional indexed data.

Versioned

An versioned tree with no additional indexed data.