Struct nebari::tree::TreeFile

pub struct TreeFile<Root: Root, File: ManagedFile> {
    pub state: State<Root>,
    // some fields omitted
}

An append-only tree file.

Generics

File: An ManagedFile implementor.

Fields
state: State<Root>

The state of the file.

Implementations

impl<Root: Root, File: ManagedFile> TreeFile<Root, File>

pub fn new(
file: <File::Manager as FileManager>::FileHandle,
state: State<Root>,
vault: Option<Arc<dyn AnyVault>>,
cache: Option<ChunkCache>
) -> Result<Self, Error>

Returns a tree as contained in file.

state should already be initialized using Self::initialize_state if the file exists.

pub fn read(
path: impl AsRef<Path>,
state: State<Root>,
context: &Context<File::Manager>,
transactions: Option<&TransactionManager<File::Manager>>
) -> Result<Self, Error>

Opens a tree file with read-only permissions.

pub fn write(
path: impl AsRef<Path>,
state: State<Root>,
context: &Context<File::Manager>,
transactions: Option<&TransactionManager<File::Manager>>
) -> Result<Self, Error>

Opens a tree file with the ability to read and write.

pub fn initialize_state(
state: &State<Root>,
file_path: &Path,
file_id: Option<u64>,
context: &Context<File::Manager>,
transaction_manager: Option<&TransactionManager<File::Manager>>
) -> Result<(), Error>

Attempts to load the last saved state of this tree into state.

pub fn push(
&mut self,
transaction_id: Option<u64>,
key: Buffer<'static>,
value: Buffer<'static>
) -> Result<(), Error>

Pushes a key/value pair. Replaces any previous value if set.

pub fn modify(
&mut self,
modification: Modification<'_, Buffer<'static>>
) -> Result<(), Error>

Executes a modification.

pub fn compare_and_swap(
&mut self,
key: &[u8],
old: Option<&Buffer<'_>>,
new: Option<Buffer<'_>>,
transaction_id: Option<u64>
) -> Result<(), CompareAndSwapError>

Compares the value of key against old. If the values match, key will be set to the new value if new is Some or removed if new is None.

pub fn remove(
&mut self,
key: &[u8],
transaction_id: Option<u64>
) -> Result<Option<Buffer<'static>>, Error>

Removes key and returns the existing value, if present.

pub fn replace(
&mut self,
key: impl Into<Buffer<'static>>,
value: impl Into<Buffer<'static>>,
transaction_id: Option<u64>
) -> Result<Option<Buffer<'static>>, Error>

Sets key to value. If a value already exists, it will be returned.

pub fn get<'k>(
&mut self,
key: &'k [u8],
in_transaction: bool
) -> Result<Option<Buffer<'static>>, Error>

Gets the value stored for key.

pub fn get_multiple(
&mut self,
keys: &[&[u8]],
in_transaction: bool
) -> Result<Vec<(Buffer<'static>, Buffer<'static>)>, Error>

Gets the values stored in keys. Does not error if a key is missing. Returns key/value pairs in an unspecified order.

pub fn get_range<'b, B: RangeBounds<Buffer<'b>> + Debug + 'static>(
&mut self,
range: B,
in_transaction: bool
) -> Result<Vec<(Buffer<'static>, Buffer<'static>)>, Error>

Retrieves all keys and values with keys that are contained by range.

pub fn scan<'range, CallerError, Range, NodeEvaluator, KeyEvaluator, DataCallback>(
&mut self,
range: Range,
forwards: bool,
in_transaction: bool,
node_evaluator: &mut NodeEvaluator,
key_evaluator: &mut KeyEvaluator,
key_reader: &mut DataCallback
) -> Result<(), AbortError<CallerError>> where
Range: RangeBounds<Buffer<'range>> + Debug + 'static,
NodeEvaluator: FnMut(&Buffer<'static>, &Root::ReducedIndex, usize) -> bool,
KeyEvaluator: FnMut(&Buffer<'static>, &Root::Index) -> KeyEvaluation,
DataCallback: FnMut(Buffer<'static>, &Root::Index, Buffer<'static>) -> Result<(), AbortError<CallerError>>,
CallerError: Display + Debug,

Scans the tree for keys that are contained within range. If forwards is true, scanning starts at the lowest sort-order key and scans forward. Otherwise, scanning starts at the highest sort-order key and scans backwards. key_evaluator is invoked for each key as it is encountered. For all KeyEvaluation::ReadData results returned, callback will be invoked with the key and values. The callback may not be invoked in the same order as the keys are scanned.

pub fn last_key(
&mut self,
in_transaction: bool
) -> Result<Option<Buffer<'static>>, Error>

Returns the last key of the tree.

pub fn last(
&mut self,
in_transaction: bool
) -> Result<Option<(Buffer<'static>, Buffer<'static>)>, Error>

Returns the last key and value of the tree.

pub fn commit(&mut self) -> Result<(), Error>

Commits the tree. This is only needed if writes were done with a transaction id. This will fully flush the tree and publish the transactional state to be available to readers.

pub fn compact(
self,
file_manager: &File::Manager,
transactions: Option<TransactableCompaction<'_, File::Manager>>
) -> Result<Self, Error>

Rewrites the database, removing all unused data in the process. For a VersionedTreeRoot, this will remove old version information.

This process is done atomically by creating a new file containing the active data. Once the new file has all the current file’s data, the file contents are swapped using atomic file operations.

impl<File: ManagedFile, Index> TreeFile<VersionedTreeRoot<Index>, File> where
Index: EmbeddedIndex + Reducer<Index> + Clone + Debug + 'static,

pub fn scan_sequences<CallerError, Range, KeyEvaluator, DataCallback>(
&mut self,
range: Range,
forwards: bool,
in_transaction: bool,
key_evaluator: &mut KeyEvaluator,
data_callback: &mut DataCallback
) -> Result<(), AbortError<CallerError>> where
Range: RangeBounds<u64> + Debug + 'static,
KeyEvaluator: FnMut(KeySequence) -> KeyEvaluation,
DataCallback: FnMut(KeySequence, Buffer<'static>) -> Result<(), AbortError<CallerError>>,
CallerError: Display + Debug,

Scans the tree for keys that are contained within range. If forwards is true, scanning starts at the lowest sort-order key and scans forward. Otherwise, scanning starts at the highest sort-order key and scans backwards. key_evaluator is invoked for each key as it is encountered. For all KeyEvaluation::ReadData results returned, callback will be invoked with the key and values. The callback may not be invoked in the same order as the keys are scanned.