Trait nebari::tree::Root

pub trait Root: Default + Debug + Send + Sync + Clone + 'static {
    const HEADER: PageHeader;

    fn count(&self) -> u64;

    fn dirty(&self) -> bool;

    fn initialized(&self) -> bool;

    fn initialize_default(&mut self);

    fn serialize<File: ManagedFile>(
        &mut self, 
        paged_writer: &mut PagedWriter<'_, File>, 
        output: &mut Vec<u8>
    ) -> Result<(), Error>;

    fn deserialize(bytes: Buffer<'_>) -> Result<Self, Error>;

    fn transaction_id(&self) -> u64;

    fn modify<'a, 'w, File: ManagedFile>(
        &'a mut self, 
        modification: Modification<'_, Buffer<'static>>, 
        writer: &'a mut PagedWriter<'w, File>
    ) -> Result<(), Error>;

    fn get_multiple<File: ManagedFile, KeyEvaluator, KeyReader>(
        &self, 
        keys: &mut KeyRange<'_>, 
        key_evaluator: &mut KeyEvaluator, 
        key_reader: &mut KeyReader, 
        file: &mut File, 
        vault: Option<&dyn Vault>, 
        cache: Option<&ChunkCache>
    ) -> Result<(), Error>
    where
        KeyEvaluator: FnMut(&Buffer<'static>) -> KeyEvaluation,
        KeyReader: FnMut(Buffer<'static>, Buffer<'static>) -> Result<(), Error>;

    fn scan<'keys, CallerError: Display + Debug, File: ManagedFile, KeyRangeBounds, KeyEvaluator, KeyReader>(
        &self, 
        range: &KeyRangeBounds, 
        args: &mut ScanArgs<Buffer<'static>, CallerError, KeyEvaluator, KeyReader>, 
        file: &mut File, 
        vault: Option<&dyn Vault>, 
        cache: Option<&ChunkCache>
    ) -> Result<(), AbortError<CallerError>>
    where
        KeyEvaluator: FnMut(&Buffer<'static>) -> KeyEvaluation,
        KeyReader: FnMut(Buffer<'static>, Buffer<'static>) -> Result<(), AbortError<CallerError>>,
        KeyRangeBounds: RangeBounds<Buffer<'keys>> + Debug;

    fn copy_data_to<File: ManagedFile>(
        &mut self, 
        include_nodes: bool, 
        file: &mut File, 
        copied_chunks: &mut HashMap<u64, u64>, 
        writer: &mut PagedWriter<'_, File>, 
        vault: Option<&dyn Vault>
    ) -> Result<(), Error>;

    fn tree<File: ManagedFile>(
        name: impl Into<Cow<'static, str>>
    ) -> TreeRoot<File> { ... }
}

A B-Tree root implementation.

Associated Constants

const HEADER: PageHeader

The unique header byte for this root.

Required methods

fn count(&self) -> u64

Returns the number of values contained in this tree, not including deleted records.

fn dirty(&self) -> bool

Returns true if the root needs to be saved.

fn initialized(&self) -> bool

Returns true if the tree is initialized.

fn initialize_default(&mut self)

Resets the state to the default, initialized state. After calling this, Self::initialized() will return true.

fn serialize<File: ManagedFile>(
&mut self,
paged_writer: &mut PagedWriter<'_, File>,
output: &mut Vec<u8>
) -> Result<(), Error>

Serialize the root and return the bytes. Writes any additional data to paged_writer in the process of serialization.

fn deserialize(bytes: Buffer<'_>) -> Result<Self, Error>

Deserializes the root from bytes.

fn transaction_id(&self) -> u64

Returns the current transaction id.

fn modify<'a, 'w, File: ManagedFile>(
&'a mut self,
modification: Modification<'_, Buffer<'static>>,
writer: &'a mut PagedWriter<'w, File>
) -> Result<(), Error>

Modifies the tree.

fn get_multiple<File: ManagedFile, KeyEvaluator, KeyReader>(
&self,
keys: &mut KeyRange<'_>,
key_evaluator: &mut KeyEvaluator,
key_reader: &mut KeyReader,
file: &mut File,
vault: Option<&dyn Vault>,
cache: Option<&ChunkCache>
) -> Result<(), Error> where
KeyEvaluator: FnMut(&Buffer<'static>) -> KeyEvaluation,
KeyReader: FnMut(Buffer<'static>, Buffer<'static>) -> Result<(), Error>,

Iterates over the tree looking for keys. keys must be sorted. key_evaluator is invoked for each key as it is found, allowing for decisions on how to handle each key. key_reader will be invoked for each key that is requested to be read, but it might be invoked at a later time and in a different order.

fn scan<'keys, CallerError: Display + Debug, File: ManagedFile, KeyRangeBounds, KeyEvaluator, KeyReader>(
&self,
range: &KeyRangeBounds,
args: &mut ScanArgs<Buffer<'static>, CallerError, KeyEvaluator, KeyReader>,
file: &mut File,
vault: Option<&dyn Vault>,
cache: Option<&ChunkCache>
) -> Result<(), AbortError<CallerError>> where
KeyEvaluator: FnMut(&Buffer<'static>) -> KeyEvaluation,
KeyReader: FnMut(Buffer<'static>, Buffer<'static>) -> Result<(), AbortError<CallerError>>,
KeyRangeBounds: RangeBounds<Buffer<'keys>> + Debug,

Scans the tree over range. args.key_evaluator is invoked for each key as it is found, allowing for decisions on how to handle each key. args.key_reader will be invoked for each key that is requested to be read, but it might be invoked at a later time and in a different order.

fn copy_data_to<File: ManagedFile>(
&mut self,
include_nodes: bool,
file: &mut File,
copied_chunks: &mut HashMap<u64, u64>,
writer: &mut PagedWriter<'_, File>,
vault: Option<&dyn Vault>
) -> Result<(), Error>

Copies all data from file into writer, updating self with the new file positions.

Provided methods

fn tree<File: ManagedFile>(name: impl Into<Cow<'static, str>>) -> TreeRoot<File>

Returns a reference to a named tree that contains this type of root.